Properly initialize window navigation history list.
When starting Vim as follows: 'vim a b c', the files a, b, and c were
added to the buffer navigation history in the wrong order. The user had
to use BufSurfBack to navigate to the buffers for files b and c. Now, a
is in the initial buffer, and issueing BufSurfForward navigates to the
buffers for file b and c respectively.